Refactor nlp.py and extra functions in flow to nlp_utils.py
